+++ packages/uploadthing/test/__test-helpers.ts
@@ -3,6 +3,7 @@ import { beforeEach, vi } from "vitest";
 import type { FetchEsque } from "@uploadthing/shared";
 
 import { UPLOADTHING_VERSION } from "../src/constants";
+import type { PSPResponse } from "../src/internal/handler";
 import type { ActionType } from "../src/server";
 
 export const fetchMock = vi.fn();
@@ -26,24 +27,58 @@ export const baseHeaders = {
   "x-uploadthing-package": "vitest",
 };
 
+const mockedPresignedPost: PSPResponse = {
+  key: "abc-123.txt",
+  url: "https://bucket.s3.amazonaws.com/abc-123.txt",
+  fields: { key: "abc-123.txt" },
+  fileUrl: "https://utfs.io/f/abc-123.txt",
+  contentDisposition: "inline",
+  fileName: "foo.txt",
+  fileType: "text/plain",
+  pollingUrl: "https://uploadthing.com/api/serverCallback",
+  pollingJwt: "random-jwt",
+};
+
 export const mockExternalRequests: FetchEsque = async (url, init) => {
   fetchMock(url, init);
   if (url instanceof Request) return new Response("Wut?", { status: 500 });
   url = new URL(url);
 
-  // If request is going to uploadthing, mock the response
+  // Mock file downloads
+  if (url.host === "cdn.foo.com") {
+    return new Response("Lorem ipsum doler sit amet", {
+      headers: { "Content-Type": "text/plain" },
+    });
+  }
+
+  // Mock UT Api
   if (url.host === "uploadthing.com") {
-    switch (url.pathname) {
-      case "/api/prepareUpload": {
-        return Response.json([]);
-      }
-      case "/api/completeMultipart": {
-        return Response.json({ success: true });
-      }
-      case "/api/failureCallback": {
-        return Response.json({ success: true });
-      }
+    if (url.pathname === "/api/uploadFiles") {
+      return Response.json({
+        // FIXME: Read body and return a better response?
+        data: [mockedPresignedPost, mockedPresignedPost],
+      });
     }
+    if (url.pathname === "/api/prepareUpload") {
+      return Response.json([mockedPresignedPost, mockedPresignedPost]);
+    }
+    if (url.pathname === "/api/completeMultipart") {
+      return Response.json({ success: true });
+    }
+    if (url.pathname === "/api/failureCallback") {
+      return Response.json({ success: true });
+    }
+    if (url.pathname.startsWith("/api/pollUpload")) {
+      return Response.json({ status: "done" });
+    }
+    if (url.pathname === "/api/requestFileAccess") {
+      return Response.json({ url: "https://example.com" });
+    }
+  }
+
+  // Mock S3
+  if (url.host === "bucket.s3.amazonaws.com") {
+    return new Response(null, { status: 204 });
   }
 
   // Else 404 (shouldn't happen, mock the requests we expect to make)
